Refactor configload files: ordered methods by context
