+++ config/configload/load.go
@@ -2,18 +2,11 @@ package configload
 
 import (
 	"fmt"
-	"net/http"
 	"os"
 	"path/filepath"
 	"regexp"
 
-	"github.com/hashicorp/hcl/v2"
-	"github.com/hashicorp/hcl/v2/gohcl"
-	"github.com/hashicorp/hcl/v2/hclsyntax"
-	"github.com/zclconf/go-cty/cty"
-
 	"github.com/avenga/couper/config"
-	hclbody "github.com/avenga/couper/config/body"
 	"github.com/avenga/couper/config/configload/collect"
 	"github.com/avenga/couper/config/parser"
 	"github.com/avenga/couper/config/reader"
@@ -21,6 +14,8 @@ import (
 	"github.com/avenga/couper/eval"
 	"github.com/avenga/couper/eval/lib"
 	"github.com/avenga/couper/internal/seetie"
+	"github.com/hashicorp/hcl/v2"
+	"github.com/hashicorp/hcl/v2/gohcl"
 )
 
 const (
@@ -313,7 +308,7 @@ func LoadConfig(body hcl.Body, src []byte, filename string) (*config.Couper, err
 				return nil, err
 			}
 
-			apiConfig.CatchAllEndpoint = createCatchAllEndpoint()
+			apiConfig.CatchAllEndpoint = newCatchAllEndpoint()
 			serverConfig.APIs = append(serverConfig.APIs, apiConfig)
 
 			apiErrorHandler := collect.ErrorHandlerSetters(apiConfig)
@@ -336,537 +331,4 @@ func LoadConfig(body hcl.Body, src []byte, filename string) (*config.Couper, err
 	}
 
 	return couperConfig, nil
-}
-
-// mergeBackendBodies appends the left side object with newly defined attributes or overrides already defined ones.
-func mergeBackendBodies(definedBackends Backends, inline config.Inline) (hcl.Body, error) {
-	var reference hcl.Body
-	if beRef, ok := inline.(config.BackendReference); ok {
-		r, err := getBackendReference(definedBackends, beRef)
-		if err != nil {
-			return nil, err
-		}
-		reference = r
-	}
-
-	content, _, diags := inline.HCLBody().PartialContent(inline.Schema(true))
-	if diags.HasErrors() {
-		return nil, diags
-	}
-
-	if content == nil {
-		if reference != nil {
-			return reference, nil
-		}
-		return nil, fmt.Errorf("configuration error: missing backend reference or inline definition")
-	}
-
-	// Apply current attributes to the referenced body.
-	if len(content.Attributes) > 0 && reference != nil {
-		reference = MergeBodies([]hcl.Body{reference, hclbody.New(&hcl.BodyContent{
-			Attributes:       content.Attributes,
-			MissingItemRange: content.MissingItemRange,
-		})})
-	}
-
-	var backendBlock *hcl.Block
-	if backends := content.Blocks.OfType(backend); len(backends) > 0 {
-		backendBlock = backends[0]
-	} else {
-		return reference, nil
-	}
-
-	// Case: `backend {}`, anonymous backend.
-	if len(backendBlock.Labels) == 0 {
-		return backendBlock.Body, nil
-	}
-
-	// Case: `backend "reference" {}`, referenced backend.
-	refOverride, err := definedBackends.WithName(backendBlock.Labels[0])
-	if err != nil {
-		err.(hcl.Diagnostics)[0].Subject = &backendBlock.DefRange
-
-		// Case: referenced backend is not defined in definitions.
-		return nil, err
-	}
-
-	// link backend block name (label) to attribute 'name'
-	if syntaxBody, ok := backendBlock.Body.(*hclsyntax.Body); ok {
-		if refBody, ok := refOverride.(*hclsyntax.Body); ok {
-			syntaxBody.Attributes[nameLabel] = refBody.Attributes[nameLabel]
-		}
-	}
-
-	return MergeBodies([]hcl.Body{refOverride, backendBlock.Body}), nil
-}
-
-// getBackendReference tries to fetch a backend from `definitions`
-// block by a reference name, e.g. `backend = "name"`.
-func getBackendReference(definedBackends Backends, be config.BackendReference) (hcl.Body, error) {
-	name := be.Reference()
-
-	// backend string attribute just not set
-	if name == "" {
-		return nil, nil
-	}
-
-	reference, err := definedBackends.WithName(name)
-	if err != nil {
-		return nil, err // parse err
-	}
-
-	// a name is given but we have no definition
-	if body, ok := be.(config.Inline); ok {
-		if b := body.HCLBody(); reference == nil && b != nil {
-			r := b.MissingItemRange()
-			return nil, hcl.Diagnostics{&hcl.Diagnostic{
-				Subject: &r,
-				Summary: fmt.Sprintf("backend reference '%s' is not defined", name),
-			}}
-		}
-	}
-
-	return reference, nil
-}
-
-func refineEndpoints(definedBackends Backends, endpoints config.Endpoints, check bool) error {
-	var err error
-
-	for _, endpoint := range endpoints {
-		if check && endpoint.Pattern == "" {
-			var r hcl.Range
-			if endpoint.Remain != nil {
-				r = endpoint.Remain.MissingItemRange()
-			}
-			return hcl.Diagnostics{&hcl.Diagnostic{
-				Severity: hcl.DiagError,
-				Summary:  "endpoint: missing path pattern",
-				Subject:  &r,
-			}}
-		}
-
-		endpointContent := bodyToContent(endpoint.Remain)
-
-		proxies := endpointContent.Blocks.OfType(proxy)
-		requests := endpointContent.Blocks.OfType(request)
-
-		if check && len(proxies)+len(requests) == 0 && endpoint.Response == nil {
-			return hcl.Diagnostics{&hcl.Diagnostic{
-				Severity: hcl.DiagError,
-				Summary:  "missing 'default' proxy or request block, or a response definition",
-				Subject:  &endpointContent.MissingItemRange,
-			}}
-		}
-
-		proxyRequestLabelRequired := len(proxies)+len(requests) > 1
-
-		for _, proxyBlock := range proxies {
-			proxyConfig := &config.Proxy{}
-			if diags := gohcl.DecodeBody(proxyBlock.Body, envContext, proxyConfig); diags.HasErrors() {
-				return diags
-			}
-			if len(proxyBlock.Labels) > 0 {
-				proxyConfig.Name = proxyBlock.Labels[0]
-			}
-			if proxyConfig.Name == "" {
-				proxyConfig.Name = defaultNameLabel
-			}
-
-			wsEnabled, wsBody, wsErr := getWebsocketsConfig(proxyConfig)
-			if wsErr != nil {
-				return wsErr
-			}
-
-			if wsEnabled {
-				if proxyConfig.Name != defaultNameLabel {
-					return errors.Configuration.Message("websockets attribute or block is only allowed in a 'default' proxy block")
-				}
-				if proxyRequestLabelRequired || endpoint.Response != nil {
-					return errors.Configuration.Message("websockets are allowed in the endpoint; other 'proxy', 'request' or 'response' blocks are not allowed")
-				}
-
-				if wsBody != nil {
-					proxyBlock.Body = MergeBodies([]hcl.Body{proxyBlock.Body, wsBody})
-				}
-			}
-
-			proxyConfig.Remain = proxyBlock.Body
-
-			proxyConfig.Backend, err = newBackend(definedBackends, proxyConfig)
-			if err != nil {
-				return err
-			}
-
-			endpoint.Proxies = append(endpoint.Proxies, proxyConfig)
-		}
-
-		for _, reqBlock := range requests {
-			reqConfig := &config.Request{}
-			if diags := gohcl.DecodeBody(reqBlock.Body, envContext, reqConfig); diags.HasErrors() {
-				return diags
-			}
-
-			if len(reqBlock.Labels) > 0 {
-				reqConfig.Name = reqBlock.Labels[0]
-			}
-			if reqConfig.Name == "" {
-				reqConfig.Name = defaultNameLabel
-			}
-
-			// remap request specific names for headers and query to well known ones
-			content, leftOvers, diags := reqBlock.Body.PartialContent(reqConfig.Schema(true))
-			if diags.HasErrors() {
-				return diags
-			}
-
-			if err := verifyBodyAttributes(content); err != nil {
-				return err
-			}
-
-			renameAttribute(content, "headers", "set_request_headers")
-			renameAttribute(content, "query_params", "set_query_params")
-
-			reqConfig.Remain = MergeBodies([]hcl.Body{leftOvers, hclbody.New(content)})
-
-			reqConfig.Backend, err = newBackend(definedBackends, reqConfig)
-			if err != nil {
-				return err
-			}
-
-			endpoint.Requests = append(endpoint.Requests, reqConfig)
-		}
-
-		if endpoint.Response != nil {
-			content, _, _ := endpoint.Response.HCLBody().PartialContent(config.ResponseInlineSchema)
-			_, existsBody := content.Attributes["body"]
-			_, existsJsonBody := content.Attributes["json_body"]
-			if existsBody && existsJsonBody {
-				return hcl.Diagnostics{&hcl.Diagnostic{
-					Severity: hcl.DiagError,
-					Summary:  "response can only have one of body or json_body attributes",
-					Subject:  &content.Attributes["body"].Range,
-				}}
-			}
-		}
-
-		names := map[string]struct{}{}
-		unique := map[string]struct{}{}
-		itemRange := endpoint.Remain.MissingItemRange()
-		for _, p := range endpoint.Proxies {
-			names[p.Name] = struct{}{}
-
-			if err := validLabelName(p.Name, &itemRange); err != nil {
-				return err
-			}
-
-			if proxyRequestLabelRequired {
-				if err := uniqueLabelName(unique, p.Name, &itemRange); err != nil {
-					return err
-				}
-			}
-		}
-
-		for _, r := range endpoint.Requests {
-			names[r.Name] = struct{}{}
-
-			if err = validLabelName(r.Name, &itemRange); err != nil {
-				return err
-			}
-
-			if proxyRequestLabelRequired {
-				if err = uniqueLabelName(unique, r.Name, &itemRange); err != nil {
-					return err
-				}
-			}
-		}
-
-		if _, ok := names[defaultNameLabel]; check && !ok && endpoint.Response == nil {
-			return hcl.Diagnostics{&hcl.Diagnostic{
-				Severity: hcl.DiagError,
-				Summary:  "Missing a 'default' proxy or request definition, or a response block",
-				Subject:  &itemRange,
-			}}
-		}
-
-		addSequenceDeps(names, endpoint)
-
-		epErrorHandler := collect.ErrorHandlerSetters(endpoint)
-		if err := configureErrorHandler(epErrorHandler, definedBackends); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func collectExpressions(bodies ...hcl.Body) []hcl.Expression {
-	allExpressions := make([]hcl.Expression, 0)
-
-	for _, b := range bodies {
-		switch sb := b.(type) {
-		case *hclsyntax.Body:
-			for _, attr := range sb.Attributes {
-				allExpressions = append(allExpressions, attr.Expr)
-			}
-		case mergedBodies:
-			for _, attrs := range sb.JustAllAttributes() {
-				for _, attr := range attrs {
-					allExpressions = append(allExpressions, attr.Expr)
-				}
-			}
-		}
-	}
-	return allExpressions
-}
-
-// addSequenceDeps collects possible dependencies from variables.
-func addSequenceDeps(names map[string]struct{}, endpoint *config.Endpoint) {
-	var items []config.SequenceItem
-	for _, conf := range endpoint.Proxies {
-		items = append(items, conf)
-	}
-	for _, conf := range endpoint.Requests {
-		items = append(items, conf)
-	}
-
-	for _, seqItem := range items {
-		allExpressions := collectExpressions(seqItem.HCLBody())
-		for _, expr := range allExpressions {
-			for _, traversal := range expr.Variables() {
-				if traversal.RootName() != "backend_responses" || len(traversal) < 2 {
-					continue
-				}
-
-				// do we have a ref ?
-				for _, t := range traversal[1:] {
-					tr, ok := t.(hcl.TraverseAttr)
-					if !ok {
-						continue
-					}
-
-					_, ok = names[tr.Name]
-					if !ok {
-						continue
-					}
-
-					for _, i := range items {
-						if i.GetName() != tr.Name || i == seqItem {
-							continue
-						}
-						seqItem.Add(i)
-						break
-					}
-				}
-			}
-		}
-	}
-}
-
-func getWebsocketsConfig(proxyConfig *config.Proxy) (bool, hcl.Body, error) {
-	content, _, diags := proxyConfig.Remain.PartialContent(
-		&hcl.BodySchema{Blocks: []hcl.BlockHeaderSchema{{Type: "websockets"}}},
-	)
-	if diags.HasErrors() {
-		return false, nil, diags
-	}
-
-	if proxyConfig.Websockets != nil && len(content.Blocks.OfType("websockets")) > 0 {
-		return false, nil, fmt.Errorf("either websockets attribute or block is allowed")
-	}
-
-	if proxyConfig.Websockets != nil {
-		var body hcl.Body
-
-		if *proxyConfig.Websockets {
-			block := &hcl.Block{
-				Type: "websockets",
-				Body: EmptyBody(),
-			}
-
-			body = hclbody.New(&hcl.BodyContent{Blocks: []*hcl.Block{block}})
-		}
-
-		return *proxyConfig.Websockets, body, nil
-	}
-
-	return len(content.Blocks) > 0, nil, nil
-}
-
-func verifyBodyAttributes(content *hcl.BodyContent) error {
-	_, existsBody := content.Attributes["body"]
-	_, existsFormBody := content.Attributes["form_body"]
-	_, existsJsonBody := content.Attributes["json_body"]
-	if existsBody && existsFormBody || existsBody && existsJsonBody || existsFormBody && existsJsonBody {
-		rangeAttr := "body"
-		if !existsBody {
-			rangeAttr = "form_body"
-		}
-		return hcl.Diagnostics{&hcl.Diagnostic{
-			Severity: hcl.DiagError,
-			Summary:  "request can only have one of body, form_body or json_body attributes",
-			Subject:  &content.Attributes[rangeAttr].Range,
-		}}
-	}
-	return nil
-}
-
-func bodyToContent(body hcl.Body) *hcl.BodyContent {
-	content := &hcl.BodyContent{
-		MissingItemRange: body.MissingItemRange(),
-	}
-	b, ok := body.(*hclsyntax.Body)
-	if !ok {
-		return content
-	}
-
-	if len(b.Attributes) > 0 {
-		content.Attributes = make(hcl.Attributes)
-	}
-	for name, attr := range b.Attributes {
-		content.Attributes[name] = &hcl.Attribute{
-			Name:      attr.Name,
-			Expr:      attr.Expr,
-			Range:     attr.Range(),
-			NameRange: attr.NameRange,
-		}
-	}
-
-	for _, block := range b.Blocks {
-		content.Blocks = append(content.Blocks, &hcl.Block{
-			Body:        block.Body,
-			DefRange:    block.DefRange(),
-			LabelRanges: block.LabelRanges,
-			Labels:      block.Labels,
-			Type:        block.Type,
-			TypeRange:   block.TypeRange,
-		})
-	}
-
-	return content
-}
-
-func contentByType(blockType string, body hcl.Body) (*hcl.BodyContent, error) {
-	headerSchema := &hcl.BodySchema{
-		Blocks: []hcl.BlockHeaderSchema{
-			{Type: blockType},
-		}}
-	content, _, diags := body.PartialContent(headerSchema)
-	if diags.HasErrors() {
-		derr := diags.Errs()[0].(*hcl.Diagnostic)
-		if derr.Summary == "Extraneous label for "+blockType { // retry with label
-			headerSchema.Blocks[0].LabelNames = []string{nameLabel}
-			content, _, diags = body.PartialContent(headerSchema)
-			if diags.HasErrors() { // due to interface nil check, do not return empty diags
-				return nil, diags
-			}
-			return content, nil
-		}
-		return nil, diags
-	}
-	return content, nil
-}
-
-func newBackend(definedBackends Backends, inlineConfig config.Inline) (hcl.Body, error) {
-	bend, err := mergeBackendBodies(definedBackends, inlineConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	if bend == nil {
-		// Create a default backend
-		bend = hclbody.New(&hcl.BodyContent{
-			Attributes: map[string]*hcl.Attribute{
-				"name": {
-					Name: "name",
-					Expr: &hclsyntax.LiteralValueExpr{
-						Val: cty.StringVal(defaultNameLabel),
-					},
-				},
-			},
-		})
-	}
-
-	oauth2Backend, err := newOAuthBackend(definedBackends, bend)
-	if err != nil {
-		return nil, err
-	}
-
-	if oauth2Backend != nil {
-		wrapped := hclbody.New(&hcl.BodyContent{Blocks: []*hcl.Block{
-			{Type: oauth2, Body: hclbody.New(&hcl.BodyContent{Blocks: []*hcl.Block{
-				{Type: backend, Body: oauth2Backend},
-			}})},
-		}})
-		bend = MergeBodies([]hcl.Body{bend, wrapped})
-	}
-
-	return bend, nil
-}
-
-func createCatchAllEndpoint() *config.Endpoint {
-	responseBody := hclbody.New(&hcl.BodyContent{
-		Attributes: map[string]*hcl.Attribute{
-			"status": {
-				Name: "status",
-				Expr: &hclsyntax.LiteralValueExpr{
-					Val: cty.NumberIntVal(http.StatusNotFound),
-				},
-			},
-		},
-	})
-
-	return &config.Endpoint{
-		Pattern: "/**",
-		Remain:  hclbody.New(&hcl.BodyContent{}),
-		Response: &config.Response{
-			Remain: responseBody,
-		},
-	}
-}
-
-func newOAuthBackend(definedBackends Backends, parent hcl.Body) (hcl.Body, error) {
-	innerContent, err := contentByType(oauth2, parent)
-	if err != nil {
-		return nil, err
-	}
-
-	oauthBlocks := innerContent.Blocks.OfType(oauth2)
-	if len(oauthBlocks) == 0 {
-		return nil, nil
-	}
-
-	backendContent, err := contentByType(backend, oauthBlocks[0].Body)
-	if err != nil {
-		return nil, err
-	}
-
-	beConfig := &config.Backend{Remain: hclbody.New(backendContent)}
-
-	attrs, _ := oauthBlocks[0].Body.JustAttributes()
-	if attrs != nil && attrs["backend"] != nil {
-		val, _ := attrs["backend"].Expr.Value(nil)
-
-		if ref := seetie.ValueToString(val); ref != "" {
-			beConfig.Name = ref
-		}
-	}
-
-	oauthBackend, err := mergeBackendBodies(definedBackends, beConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	return newBackend(definedBackends, &config.OAuth2ReqAuth{Remain: hclbody.New(&hcl.BodyContent{
-		Blocks: []*hcl.Block{
-			{Type: backend, Body: oauthBackend},
-		},
-	})})
-}
-
-func renameAttribute(content *hcl.BodyContent, old, new string) {
-	if attr, ok := content.Attributes[old]; ok {
-		attr.Name = new
-		content.Attributes[new] = attr
-		delete(content.Attributes, old)
-	}
 }
