+++ config/configload/validate.go
@@ -35,3 +35,21 @@ func uniqueLabelName(unique map[string]struct{}, name string, hr *hcl.Range) err
 	unique[name] = struct{}{}
 	return nil
 }
+
+func verifyBodyAttributes(content *hcl.BodyContent) error {
+	_, existsBody := content.Attributes["body"]
+	_, existsFormBody := content.Attributes["form_body"]
+	_, existsJsonBody := content.Attributes["json_body"]
+	if existsBody && existsFormBody || existsBody && existsJsonBody || existsFormBody && existsJsonBody {
+		rangeAttr := "body"
+		if !existsBody {
+			rangeAttr = "form_body"
+		}
+		return hcl.Diagnostics{&hcl.Diagnostic{
+			Severity: hcl.DiagError,
+			Summary:  "request can only have one of body, form_body or json_body attributes",
+			Subject:  &content.Attributes[rangeAttr].Range,
+		}}
+	}
+	return nil
+}
